+++ rust/src/pr020.rs
@@ -19,24 +19,24 @@ use std::iter;
 
 define_problem!(pr020, 20, 648);
 
-static BASE: uint = 10_000;
+static BASE: u64 = 10_000;
 
-fn pr020() -> uint {
-    static SIZE: uint = 40;
-    let mut acc: Vec<_> = iter::repeat(0u).take(SIZE).collect();
+fn pr020() -> u64 {
+    static SIZE: usize = 40;
+    let mut acc: Vec<_> = iter::repeat(0).take(SIZE).collect();
 
     acc[0] = 1;
 
-    for x in 2u .. 101 {
+    for x in 2 .. 101 {
         multiply(acc.as_mut_slice(), x);
     }
     sum_digits(&acc[..])
 }
 
 // Multiply the little-endian base 10_000 number in 'acc' by 'by'.
-fn multiply(acc: &mut [uint], by: uint) {
-    let mut carry = 0u;
-    for i in 0u .. acc.len() {
+fn multiply(acc: &mut [u64], by: u64) {
+    let mut carry = 0;
+    for i in 0 .. acc.len() {
         let temp = acc[i] * by + carry;
         acc[i] = temp % BASE;
         carry = temp / BASE;
@@ -44,9 +44,9 @@ fn multiply(acc: &mut [uint], by: uint) {
     if carry != 0 { panic!("Multiply overflow") }
 }
 
-fn sum_digits(acc: &[uint]) -> uint {
-    acc.iter().fold(0u, |accum, &n| {
-        let mut sub = 0u;
+fn sum_digits(acc: &[u64]) -> u64 {
+    acc.iter().fold(0, |accum, &n| {
+        let mut sub = 0;
         let mut tmp = n;
         while tmp > 0 {
             sub += tmp % 10;
