+++ rust/src/pr031.rs
@@ -20,12 +20,12 @@
 
 define_problem!(pr031, 31, 73682);
 
-fn pr031() -> uint {
+fn pr031() -> u64 {
     let coins = all_coins();
     rways(200, &coins[..])
 }
 
-fn rways(remaining: uint, coins: &[uint]) -> uint {
+fn rways(remaining: u64, coins: &[u64]) -> u64 {
     if coins.is_empty() {
         if remaining == 0 { 1 } else { 0 }
     } else {
@@ -33,17 +33,17 @@ fn rways(remaining: uint, coins: &[uint]) -> uint {
         let others = coins.tail();
 
         let mut total = 0;
-        let mut r = remaining as int;
+        let mut r = remaining as i64;
         while r >= 0 {
-            let r2 = r as uint;
+            let r2 = r as u64;
             total += rways(r2, others);
 
-            r -= coin as int;
+            r -= coin as i64;
         }
         total
     }
 }
 
-fn all_coins() -> Vec<uint> {
+fn all_coins() -> Vec<u64> {
     vec![200, 100, 50, 20, 10, 5, 2, 1]
 }
